Improve parsing
The new technique improves clarity and conciseness and avoids using any external dependencies like Logos.
It also makes the `{...}` format more flexible, so any affix, like `-|` or `&`, can be used as a prefix or suffix within the brackets, in addition to by itself.

I also made pointless brackets a hard error and fixed the entries in the dictionary that triggered the error.