Add "reset" command
This is useful if you switch to another input box and want to reset the engine to its initial state when starting to write something new. Combined with SP and KPA you can get to any needed input state this way.

Note that in the long run it would be better to be able to remember multiple states simultaneously. But this is out of scope from Steno itself because it can simply be simulated by having multiple Stenos.