Fix representation of suffixes in backlog
Well, I broke the cardinal rule of the backlog, again, but in a different way. Because, for suffixes, I grouped their strokes in with the entry the suffix was added to, if the suffix stroke was a separate stroke, it was not properly considered as separate by the multi-stroke resolution algorithm.

The solution is to leave the suffix stroke as a separate entry. However, how can the suffix stroke overwrite the contents appended by a previous entry? To solve this, we add a "replaces previous" field to the backlog entry, so that when an entry with this flag is deleted (and some other checks are satisfied; see the diff), the contents of the previous entry, which were deleted when this entry was appended, are _re-_appended.